// utils/api/generated/models/view-all-sub-departments200-response-data-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Chatbot
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ViewAllSubDepartments200ResponseDataInner
 */
export interface ViewAllSubDepartments200ResponseDataInner {
  /**
   *
   * @type {string}
   * @memberof ViewAllSubDepartments200ResponseDataInner
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof ViewAllSubDepartments200ResponseDataInner
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof ViewAllSubDepartments200ResponseDataInner
   */
  visibility: string;
  /**
   *
   * @type {Array<string>}
   * @memberof ViewAllSubDepartments200ResponseDataInner
   */
  questions: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof ViewAllSubDepartments200ResponseDataInner
   */
  knowledgeChunks: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof ViewAllSubDepartments200ResponseDataInner
   */
  subDepartments: Array<string>;
}